requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
anyio==4.9.0
asttokens==2.4.0
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.2.0
click==8.2.1
colorama==0.4.6
colorlover==0.3.0
comm==0.1.4
contourpy==1.1.1
cufflinks==0.17.3
curl_cffi==0.10.0
cycler==0.11.0
dateparser==1.2.1
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
folium==0.19.5
fonttools==4.42.1
frozendict==2.4.6
frozenlist==1.6.0
gast==0.6.0
gunicorn
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
greenlet==3.2.2
grpcio==1.71.0
gspread==6.2.1
h11==0.16.0
h5py==3.13.0
html2image==2.0.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.2
idna==3.4
imageio==2.37.0
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.1
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
libclang==18.1.1
lxml==4.9.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
ml_dtypes==0.2.0
mpmath==1.3.0
multidict==6.4.3
multitasking==0.0.11
namex==0.0.9
nest-asyncio==1.5.8
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.3.1
obspy==1.4.2
opt_einsum==3.4.0
optree==0.15.0
packaging==23.1
pandas==2.1.1
pandas-datareader==0.10.0
parso==0.8.3
peewee==3.18.1
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.39
propcache==0.3.1
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.16.1
pyparsing==3.1.1
python-binance==1.0.28
python-dateutil==2.8.2
python-telegram-bot==20.6
python-dotenv
pytz>=2020.1
pyzmq==25.1.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==1.4.5
stack-data==0.6.2
sympy==1.12
ta==0.11.0
tenacity==8.2.3
tensorboard==2.15.1
tensorflow==2.15.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tornado==6.3.3
tradingview-ta==3.3.0
traitlets==5.10.0
typing_extensions==4.13.2
tzdata==2023.3
tzlocal==5.3.1
urllib3==2.0.5
wcwidth==0.2.6
websocket-client==1.8.0
websockets==8.1
Werkzeug==3.1.3
widgetsnbextension==4.0.9
wrapt==1.14.1
xyzservices==2025.4.0
yarl==1.20.0
yfinance==0.2.58
